Simplify StopFailure hook tests
